Move middleware tests to their own test file
